// web/src/app/api/model/dashboard-counts.model.ts
/**
 * Property Management System
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DashboardCounts { 
    propertiesTotal: number;
    propertiesActive: number;
    propertiesArchived: number;
    attachmentsTotal: number;
    /**
     * Attachments whose extraction state is `pending`.
     */
    attachmentsPending: number;
    customFieldsTotal: number;
    propertyTypesTotal: number;
    areasTotal: number;
}